test(unit): converted the testing prompt tests to vitest
for #452
